requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
qdrant-client>=1.7.0 